% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgin.R
\name{dgin}
\alias{dgin}
\title{Title General density with tau from (0, infinity)}
\usage{
dgin(z, a, m, t, log = TRUE, quasi = FALSE)
}
\arguments{
\item{z}{real number (Distribution is supported on real line, z can be any real number.Currently, only scalar z is supported. Density is set to be 0 at z = 0.)}

\item{a}{a degrees-of-freedom parameter}

\item{m}{similar to a location parameter, it shifts the density of the distribution left and right}

\item{t}{real number > 0 (Similar to scale parameter, controls spread of the distribution.)}

\item{log}{boolean, optional(Should the log of the density be returned? Default is True.)}

\item{quasi}{boolean, optional(The quasi-density or kernel is the density without the normalization constant.
Should the quasi-density value be returned? Default is False.)}
}
\value{
dgin
}
\description{
Title General density with tau from (0, infinity)
}
