% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgin1.R
\name{dgin1}
\alias{dgin1}
\title{Generalized inverse normal (log) density
Standardized density setting tau = 1
Internal function}
\usage{
dgin1(z, alpha, mu, log = TRUE, quasi = FALSE)
}
\arguments{
\item{z}{quantile.}

\item{alpha}{degrees-of-freedom parameter.}

\item{mu}{similar to location parameter, controls asymmetry of the distribution.}

\item{log}{logical; should the log of the density be returned? Defaults to TRUE.}

\item{quasi}{logical; should the quasi-density value be returned? Defaults to FALSE.}
}
\value{
dgin1
}
\description{
Generalized inverse normal (log) density
Standardized density setting tau = 1
Internal function
}
