% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtgin.R
\name{dtgin}
\alias{dtgin}
\title{Title}
\usage{
dtgin(z, a, m, t, sign = TRUE, log = TRUE, quasi = FALSE)
}
\arguments{
\item{z}{real number (Distribution is supported on real line, z can be any real number.
Currently, only scalar z is supported. Density is set to be 0 at z = 0.)}

\item{a}{real number > 1 (Degrees of freedom parameter. Function returns 0 density if a <= 1.)}

\item{m}{similar to a location parameter, it shifts the density of the distribution left and right}

\item{t}{similar to a scale parameter, it spreads the density of the distribution}

\item{sign}{sign = TRUE implies truncation to positive numbers and sign = FALSE to negative numbers}

\item{log}{should the logarithm of the density be returned? Defaults to TRUE.}

\item{quasi}{should the value of the kernel (or quasi-density) be returned? Defaults to FALSE}
}
\value{
dtgin
}
\description{
Title
}
