% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgin.R
\name{rgin}
\alias{rgin}
\title{Title Generator of generalized inverse normal variables}
\usage{
rgin(size, a, m, t, algo = TRUE)
}
\arguments{
\item{size}{integer (Number of desired draws. Output is numpy vector of length equal to size.)}

\item{a}{a degrees-of-freedom parameter,}

\item{m}{similar to a location parameter, it shifts the density of the distribution left and right}

\item{t}{similar to a scale parameter, it spreads the density of the distribution}

\item{algo}{They take an additional argument algo, which can be either "hormann" or "leydold", and defaults to "hormann"}
}
\value{
rgin
}
\description{
Title Generator of generalized inverse normal variables
}
