% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtgin.R
\name{dtgin}
\alias{dtgin}
\title{Density for the generalized inverse normal distribution truncated to the positive or negative reals}
\usage{
dtgin(z, alpha, mu, tau, sign = TRUE, log = TRUE, quasi = FALSE)
}
\arguments{
\item{z}{quantile.}

\item{alpha}{degrees-of-freedom parameter.}

\item{mu}{similar to location parameter, controls asymmetry of the distribution.}

\item{tau}{similar to scale parameter, controls spread of the distribution.}

\item{sign}{logical. `TRUE` implies truncation to positive numbers (`z` > 0)
and `FALSE` to negative numbers (`z` < 0)}

\item{log}{logical; should the log of the density be returned? Defaults to TRUE.}

\item{quasi}{logical; should the quasi-density value be returned? Defaults to FALSE.}
}
\value{
Numeric scalar with density.
}
\description{
Density for the generalized inverse normal distribution truncated to the positive or negative reals
}
\details{
Currently, only scalars are supported for the quantile and parameter values.
Density is supported on the positive reals (`z` > 0) when `sign = TRUE` and to
negative reals (`z` < 0) when `sign = FALSE`. `mu` can take any value
in $(-\\infty, \\infty)$. Density is only defined for parameter values
`alpha`${} > 1$ or `tau`${} > 0$, so it is set to 0 outside of these values.
The quasi-density or kernel is the density without the normalization constant,
use `quasi = TRUE` for this behavior.
}
