% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtgin.R
\name{rtgin}
\alias{rtgin}
\title{Title}
\usage{
rtgin(size, a, m, t, sign = TRUE, algo = TRUE, verbose = FALSE)
}
\arguments{
\item{size}{integer
(Number of desired draws. Output is numpy vector of length equal to size)}

\item{a}{real number > 2
(Degrees of freedom parameter. Currently, only values of a > 2 are supported.m : real number)}

\item{m}{real number (Similar to location parameter, controls asymmetry of the distribution.)}

\item{t}{similar to a scale parameter, it spreads the density of the distribution}

\item{sign}{where sign = TRUE implies truncation to positive numbers and sign = FALSE to negative numbers}

\item{algo}{boolean, optional
algo = True, use the Minimal Bounding Rectangle from Hörmann and Leydold (2014)
algo = False, use the one from Leydold (2001)
Defaults to True.}

\item{verbose}{boolean, optional
Should the acceptance rate from the ratio-of-uniforms method be provided
along with additional information? Defaults to False.}
}
\value{
rtgin
}
\description{
Title
}
